refactor: move Cover next to the cmponent using it
